requirements.txt
Flask==3.0.3
Flask-CORS==4.0.0
python-dotenv==1.0.1
SQLAlchemy==2.0.36
psycopg2-binary==2.9.12
requests==2.31.0
ebaysdk==2.2.0
google-auth==2.28.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.1.1
google-api-python-client==2.118.0
selenium==4.18.1
webdriver-manager==4.0.1
APScheduler==3.10.4
urllib3==2.2.1
anthropic==0.96.0
beautifulsoup4==4.14.3
cryptography>=42.0
pytest>=8.0
PyYAML>=6.0